azure_func/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions==1.18.0
azure-storage-blob==12.20.0
beautifulsoup4==4.12.3
requests==2.32.3
pandas==2.2.3